feature/implementation-services

+++ src/main/java/org/alphatrack/screensociety/repositories/contracts/CommentRepository.java
@@ -3,6 +3,6 @@ package org.alphatrack.screensociety.repositories.contracts;
 import org.alphatrack.screensociety.models.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CommentRepository extends JpaRepository<Comment, Integer>,CommentRepositoryCustom {
+public interface CommentRepository extends JpaRepository<Comment, Long>,CommentRepositoryCustom {
 
 }
