+++ src/main/java/org/alphatrack/screensociety/services/UserServiceImpl.java
@@ -71,7 +71,7 @@ public class UserServiceImpl implements UserService {
     @PreAuthorize("hasRole('ADMIN')")
     @Transactional
     @Override
-    public void blockUser(Integer userId) {
+    public void blockUser(Long userId) {
         User user = getUserById(userId);
 
         if (user.getRole() == Role.ADMIN) {
@@ -85,7 +85,7 @@ public class UserServiceImpl implements UserService {
     @PreAuthorize("hasRole('ADMIN')")
     @Transactional
     @Override
-    public void unBlockUser(Integer userId) {
+    public void unBlockUser(Long userId) {
         User user = getUserById(userId);
         user.setBlocked(false);
         userRepository.save(user);
@@ -94,14 +94,14 @@ public class UserServiceImpl implements UserService {
     @PreAuthorize("hasRole('ADMIN')")
     @Transactional
     @Override
-    public void promoteToAdmin(Integer userId) {
+    public void promoteToAdmin(Long userId) {
         User user = getUserById(userId);
         user.setRole(Role.ADMIN);
         userRepository.save(user);
     }
 
     @Override
-    public User getUserById(Integer id) {
+    public User getUserById(Long id) {
         return userRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException(String.format("User with id %d not found", id)));
     }
